Unity: Introduce Arduino handling script
Introduce all the UI handling for connection, deconnection and transfering data.
Implement the base logic of the communication protocol, but not the communication
itself.
This is left as an exercise for the students, following the specification given.

If successful, this should be enough to load and display a PICO-8 cartridge,
and then launch it with the PICO-8.