Arduino: Add the script to read and transmit a cart
Add the Arduino code skeleton to read from the EEPROM with I²C, checksum the data
and send it all to Unity.

All communication is removed and is expected to be implemented by the students,
following the protocol.
Constants related to the EEPROM are also removed and can be found in the datasheet.